augment log records with hostname and other runtime details
Performs a lookup of all sources on startup, and again every 5 minutes. If
a log record contains an muid, and we can resolve that muid from the
loaded sources list, the record will be augmented with a runtime_details
object containing the hostname, IP addresses, MAC addresses, and cloud
instance ID (if applicable).

Source info comes from snapshot records, which currently arrive every 15
minutes. These are ingested through the analytics platform and stored by
the API service. If any runtime details (such as IPs, hostname) change
on the host, the details applied to the log record may reflect a former
state of the host.